Create refresh_database2.py
This script is used to collect information about UFC events and fights from the UFC Stats website and store it in a SQL database. The main functionalities of the script are:

1. Connect to a local SQL database.
2. Scrape the UFC Stats website for a list of all completed events and their details, such as event ID, date, name, and location.
3. Compare the events in the database with those on the website and identify new events that have not yet been added to the database.
4. Insert new events into the SQL database.
5. Scrape the details of individual fights from each new event, such as fight ID, weight class, method, round, fight time, and winner.
6. Insert the fight details into the SQL database.

The script utilizes Selenium to navigate and interact with the UFC Stats website, and Pyodbc to interact with the SQL database. It also includes various helper functions to handle and clean the scraped data before inserting it into the database.

Is important to know that the tables already should be created into the SQL. For creating the tables into the SQL server we use a separetad script named: create_tables.